simplify bid flow setup
