+++ src/lib/Components/Bidding/Screens/SelectMaxBid.tsx
@@ -1,7 +1,7 @@
 import NavigatorIOS from "lib/utils/__legacy_do_not_use__navigator-ios-shim"
 import React from "react"
 import { ViewProperties } from "react-native"
-import { createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
+import { createRefetchContainer, graphql, QueryRenderer, RelayRefetchProp } from "react-relay"
 
 import Spinner from "../../../Components/Spinner"
 import { Schema, screenTrack } from "../../../utils/track"
@@ -16,6 +16,9 @@ import { ConfirmBidScreen } from "./ConfirmBid"
 
 import { SelectMaxBid_me } from "__generated__/SelectMaxBid_me.graphql"
 import { SelectMaxBid_sale_artwork } from "__generated__/SelectMaxBid_sale_artwork.graphql"
+import { SelectMaxBidQuery } from "__generated__/SelectMaxBidQuery.graphql"
+import { defaultEnvironment } from "lib/relay/createEnvironment"
+import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 
 interface SelectMaxBidProps extends ViewProperties {
   sale_artwork: SelectMaxBid_sale_artwork
@@ -33,7 +36,7 @@ interface SelectMaxBidState {
   context_screen: Schema.PageNames.BidFlowMaxBidPage,
   context_screen_owner_type: null,
 })
-export class SelectMaxBid extends React.Component<SelectMaxBidProps, SelectMaxBidState> {
+class SelectMaxBid extends React.Component<SelectMaxBidProps, SelectMaxBidState> {
   state = {
     selectedBidIndex: 0,
     isRefreshingSaleArtwork: false,
@@ -94,7 +97,7 @@ export class SelectMaxBid extends React.Component<SelectMaxBidProps, SelectMaxBi
   }
 }
 
-export const MaxBidScreen = createRefetchContainer(
+const SelectMaxBidContainer = createRefetchContainer(
   SelectMaxBid,
   {
     sale_artwork: graphql`
@@ -121,3 +124,38 @@ export const MaxBidScreen = createRefetchContainer(
     }
   `
 )
+
+export const SelectMaxBidQueryRenderer: React.FC<{
+  artworkID: string
+  saleID: string
+  navigator: NavigatorIOS
+}> = ({ artworkID, saleID, navigator }) => {
+  // TODO: artworkID can be nil, so omit that part of the query if it is.
+  return (
+    <>
+      <QueryRenderer<SelectMaxBidQuery>
+        environment={defaultEnvironment}
+        query={graphql`
+          query SelectMaxBidQuery($artworkID: String!, $saleID: String!) {
+            artwork(id: $artworkID) {
+              sale_artwork: saleArtwork(saleID: $saleID) {
+                ...SelectMaxBid_sale_artwork
+              }
+            }
+            me {
+              ...SelectMaxBid_me
+            }
+          }
+        `}
+        cacheConfig={{ force: true }} // We want to always fetch latest bid increments.
+        variables={{
+          artworkID,
+          saleID,
+        }}
+        render={renderWithLoadProgress<SelectMaxBidQuery["response"]>((props) => (
+          <SelectMaxBidContainer me={props.me!} sale_artwork={props.artwork!.sale_artwork!} navigator={navigator} />
+        ))}
+      />
+    </>
+  )
+}
